Add simple virtual machine.
In a new file, virtual-machine.js, a new
VirtualMachine type is defined.

It can execute simple assembly-like instructions:
        Load/Load-Immediate
        Store/Store-Immediate
        Add, Subtract, Multiply, and Divide

It has a bank of registers (unlimited in number),
and a block of memory.
Both of these things are represented as arrays.

Instructions can be given to the machine in the format
{opcode: <op>, reg: <reg>, mem: <mem>, ...}
where the exact configuration of registers and memory locations
varies according to the specific instruction.
An instruction with opcode: 'stop' will terminate the execution
of an array of instructions, at which point the run method
returns with the value of the 0-register.
The machine retains state (registers and memory, as well as
heap and stack pointers), after this termination, though.

Memory can be allocated using an allocate method.
It increments the heap pointer, and returns the location
of the allocated memory for later reference.

TODO:
        Implement stack operations (push/pop etc.)
        Implement logical tests for branching and looping
        Implement jumps for branching and function calling