Add local gitignore.
This is to ignore the .tmp.html file
used for local viewing of the index.php file.
When not using a server, to view index.php
I first run it through the php interpreter
and save the resulting html to .tmp.html.
Then, I can view .tmp.html using Firefox.